+++ app/src/main/res/layout/fragment_how_to_play_3.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/background"
     android:orientation="vertical"
-    tools:context=".how.HowToPlayFragment">
+    tools:context=".instructions.HowToPlayFragment">
 
 
     <ImageView
